New feature : Filter engine for movies list
It's now possible from the movie list to filter on :
 - Origin
 - Type
 - Movies seen on theater by a user

It has been necesseary to update the sort algorithm in order to only display
movies with a mark for a user when we select a specific user but keeping the
filter on movies seen on theater. It has been necessary to build a complex SQL
request with a subquery.