New Feature : Homework for users
Users can now give homeworks to others users meaning that they ask to another
user to rate a movie.

The homework thing is done on the mark object : An homework is a mark (movie_id
+ user_id) without mark and whithout comment but with a creation date for the
homework and the user id who added the homework.

A homework is complete when the mark has been added.

The mark object has been updated : Two fields have been added
 - homework_when : When the homework has been created
 - homework_who : Who created the homework.

movies_list template and view have been changed in order to handle homeworks.

A mail notification is now sent when an homework is created.