Fix integrity error on marks considering homeworks.
Now he have the homeworks feature, mark rows are created when homework is asked.
So, we can have an integrity error when a movie is rated. When the movie is
rated, the mark record can already exists if an homework has been created. So we
must check if the record exists before creating it. If it exists, we do a
simple update and if it doesn't exist, we create the record.