New feature : Dashboard first version
On login, we display now the last activity done on the databases by all users.
Now, we just display when movies have been added and when movies have been rated by users.

For that, datetime information have been added on the model in order to timestamp transactions on movies and marks.