New feature: Update movie data, origin and type.
It can happen that movie data are incorrect especially after a first automatic import.
This commit allows to update movies data.

An update button has been added on the movie show page which will display the
add movie wizard customized in order to execute an update in stead of an add.