Date change on a mark update has been disabled.
In order to avoid misunderstanding in the activity flow.

If we update the date in such a way, the row with the added movie event is
removed and is replaced by the updated record. It's due to the fact that we have
only one date field in the mark movie for handling that.

Closes #12