New feature : Global activity flow (Issue: #6)
This commit adds the global activity flow which allows the user to show the
whole database activity.

The activity flow view for the dashboard has been updated in order to use the
new get_activity_list function.

Navbar has been updated in order to add the link to the global activity view.