Make movie unicity contstraint on tmvdb id
Before that commit, a movie couldn't have the same title which can be bad in some cases.
Now the unicity constraint is based on the tmvdb id which that means that two
movies can't be linked to the same TMVDB ID but they can have the same title.

This commit closes #17.