Fetch all results provided by the API on a search query
The search query results are splitted on different pages with 20 records on each page.
That means if a search query returned more than 20 results, we couldn't choose
the correct movie because only the first 20 movies are available.

The new code fetch the total page number and fetch all movies into them.

This commit closes #19