Rebrand login page without modal box.
The login page has been fully rebranded in order to remove that modal box which
was difficult to handle. Then, the TMDB logo has been added in order to be
compliant with the TMDB terms of use as described on issue #23.

Thanks the autofocus parameter, the focus is now set directly on the user field.