New navigation system for graphs.
It is now possible to navigate between the graphs using a navigation bar with Prev and Next buttons.

For that navbar implementation, the graph list order has been implemented into a
app context variable and used for the navigation into the graph page but also in
the main navigation bar. Then if we add a graph on that list, it is directly
handled by the both navigation systems for graphs.

This commit closes #2.