Recode/reorganize download poster process
The download_poster function now only fetch the poster considering an URL passed
as a parameter. We don't use the movie object anymore.

The URL is now update in the get_movie function and not in the download_poster
one. This URL is updated considering the fetch_poster parameter. If it's true,
we store the local URL in the app and if it's false, we store the remote URL
from TMVDB.