Add new email notification when a comment is posted (#31)
When a new comment is posted on a mark, a mail notification is sent to all the
users if they ask for it.

This implied a profile modification : A new notification preference has been
added to handle these notifications. User model has been updated for handling
these notifications.

Comment notification handles the own activity profile option.